apps/text2img/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==1.40.0
Werkzeug==3.0.3
requests==2.31.0
